backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.1"
description = ""

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.3"
Flask-SQLAlchemy = "^2.5.1"
Flask-RESTful = "^0.3.9"
SQLAlchemy-Utils = "^0.38.2"
httpie = "^2.6.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.27.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
